+++ CHANGELOG.md
@@ -2,8 +2,22 @@
 
 ## v5.2.0-xx - xxxx-xx-xx
 
+### New features:
 
+* Added synchronisation of marketIDs with Raven Colonial now it supports them
+* Better edit validation
+* In progress builds can be deleted from Raven Colonial
+* Added Surface/Orbital indicator to Colonisation window
+* Added synchronous update when Refresh button is pressed
+* Added recording of final build costs for bases
+* Added removal of planned builds that have been removed in RC.
 
+### Bug fixes:
+
+* Fixed bug with commodity display for non-fleetcarrier owners
+* Fixed issue with a long translation string
+* Fixed bug with deactivating the first system and removing tracking of hidden systems
+* Fixed repetetive update bug with empty values from Raven Colonial
 
 ## v5.2.0-b1 - 2025-09-29
 
@@ -23,6 +37,17 @@
 
 * `/activites` endpoint: `influence` percentage added to faction information for each faction.
 
+### New features:
+
+* Colonisation: Added insert & remove menu options for adding/deleting builds
+
+### Bug fixes:
+
+* Fixed fleet carrier cargo bug
+* Improved migration of data
+* Fixed an overlay runtime error
+* Swapped order of planned and completed to be more logical
+* Small fixes for creating plans when logged out
 
 ## v5.2.0-a2 - 2025-09-20